SonarQube Fix: Resolve duplicate code by creating singular templateMixin to handle Filter Table column sorting
