+++ resources/js/processes/screen-templates/components/MyTemplatesListing.vue
@@ -138,6 +138,7 @@ import FilterTableBodyMixin from "../../../components/shared/FilterTableBodyMixi
 import paginationTable from "../../../components/shared/PaginationTable.vue";
 import fieldsMixin from "../mixins/fieldsMixin";
 import navigationMixin from "../mixins/navigationMixin";
+import templateMixin from "../mixins/templateMixin.js";
 
 const uniqIdsMixin = createUniqIdsMixin();
 
@@ -151,6 +152,7 @@ export default {
     FilterTableBodyMixin,
     fieldsMixin,
     navigationMixin,
+    templateMixin,
   ],
   props: {
     permission: {
